requirements.txt
Django
djangorestframework
drf-yasg
django_filter